minor refactoring for github comments
